Refactor DB initialization code
This simplifies the Slick configuration process and allows more code to
be shared between test and production code.
