+++ persistence-jdbc/core/src/test/scala/com/lightbend/lagom/internal/persistence/jdbc/SlickDbTestProvider.scala
@@ -3,9 +3,6 @@
  */
 package com.lightbend.lagom.internal.persistence.jdbc
 
-import javax.naming.InitialContext
-
-import com.typesafe.config.ConfigFactory
 import play.api.db.Databases
 import play.api.inject.ApplicationLifecycle
 
@@ -17,35 +14,20 @@ object SlickDbTestProvider {
   private val JNDIName = "DefaultDS"
   private val JNDIDBName = "DefaultDB"
 
+  private val AsyncExecConfig = new AsyncExecutorConfig {
+    override val numThreads: Int = 20
+    override val minConnections: Int = 20
+    override val maxConnections: Int = 100
+    override val queueSize: Int = 100
+  }
+
   /** Builds Slick Database (with AsyncExecutor) and bind it as JNDI resource for test purposes  */
   def buildAndBindSlickDb(baseName: String, lifecycle: ApplicationLifecycle): Unit = {
     val dbName = s"${baseName}_${Random.alphanumeric.take(8).mkString}"
     val db = Databases.inMemory(dbName, config = Map("jndiName" -> JNDIName))
     lifecycle.addStopHook(() => Future.successful(db.shutdown()))
 
-    val slickDb =
-      SlickDbProvider(
-        db.dataSource,
-        ConfigFactory.parseString(
-          """
-            |{
-            |  queueSize = 100
-            |  numThreads = 20
-            |  minConnections = 20
-            |  maxConnections = 100
-            |}
-          """.stripMargin
-        )
-      )
-
-    val context = new InitialContext()
-
-    context.bind(JNDIDBName, slickDb)
-    // Unbind it again when the test tears down
-    lifecycle.addStopHook { () =>
-      context.unbind(JNDIDBName)
-      slickDb.shutdown
-    }
+    SlickDbProvider.buildAndBindSlickDatabase(db, AsyncExecConfig, JNDIDBName, lifecycle)
   }
 
 }
