+++ persistence-jdbc/javadsl/src/main/scala/com/lightbend/lagom/javadsl/persistence/jdbc/JdbcPersistenceModule.scala
@@ -8,7 +8,7 @@ import javax.inject.{ Inject, Singleton }
 import akka.actor.ActorSystem
 import com.google.inject.{ AbstractModule, Key, Provider }
 import com.lightbend.lagom.internal.javadsl.persistence.jdbc._
-import com.lightbend.lagom.internal.persistence.jdbc.{ JndiConfigurator, SlickOffsetStore }
+import com.lightbend.lagom.internal.persistence.jdbc.{ SlickDbProvider, SlickOffsetStore }
 import com.lightbend.lagom.javadsl.persistence.PersistentEntityRegistry
 import com.lightbend.lagom.spi.persistence.OffsetStore
 import play.api.db.DBApi
@@ -34,7 +34,7 @@ class GuiceSlickProvider @Inject() (dbApi: DBApi, actorSystem: ActorSystem, appl
 
   lazy val get = {
     // Ensures JNDI bindings are made before we build the SlickProvider
-    JndiConfigurator(dbApi, actorSystem.settings.config, applicationLifecycle)
+    SlickDbProvider.buildAndBindSlickDatabases(dbApi, actorSystem.settings.config, applicationLifecycle)
     new SlickProvider(actorSystem)
   }
 }
