+++ persistence-jdbc/scaladsl/src/main/scala/com/lightbend/lagom/scaladsl/persistence/jdbc/JdbcPersistenceComponents.scala
@@ -3,19 +3,14 @@
  */
 package com.lightbend.lagom.scaladsl.persistence.jdbc
 
-import scala.concurrent.ExecutionContext
 import akka.actor.ActorSystem
-import com.lightbend.lagom.internal.persistence.jdbc.{ JndiConfigurator, SlickOffsetStore, SlickProvider }
-import com.lightbend.lagom.internal.scaladsl.persistence.jdbc.JdbcPersistentEntityRegistry
-import com.lightbend.lagom.internal.scaladsl.persistence.jdbc.JdbcReadSideImpl
-import com.lightbend.lagom.internal.scaladsl.persistence.jdbc.JdbcSessionImpl
-import com.lightbend.lagom.internal.scaladsl.persistence.jdbc.OffsetTableConfiguration
-import com.lightbend.lagom.scaladsl.persistence.PersistenceComponents
-import com.lightbend.lagom.scaladsl.persistence.PersistentEntityRegistry
-import com.lightbend.lagom.scaladsl.persistence.ReadSidePersistenceComponents
-import com.lightbend.lagom.scaladsl.persistence.WriteSidePersistenceComponents
+import com.lightbend.lagom.internal.persistence.jdbc.{ SlickDbProvider, SlickOffsetStore, SlickProvider }
+import com.lightbend.lagom.internal.scaladsl.persistence.jdbc.{ JdbcPersistentEntityRegistry, JdbcReadSideImpl, JdbcSessionImpl, OffsetTableConfiguration }
+import com.lightbend.lagom.scaladsl.persistence.{ PersistenceComponents, PersistentEntityRegistry, ReadSidePersistenceComponents, WriteSidePersistenceComponents }
 import com.lightbend.lagom.spi.persistence.OffsetStore
 import play.api.db.DBComponents
+
+import scala.concurrent.ExecutionContext
 
 /**
  * Persistence JDBC components (for compile-time injection).
@@ -32,7 +27,7 @@ private[lagom] trait SlickProviderComponents extends DBComponents {
 
   lazy val slickProvider: SlickProvider = {
     // Ensures JNDI bindings are made before we build the SlickProvider
-    JndiConfigurator(dbApi, actorSystem.settings.config, applicationLifecycle)
+    SlickDbProvider.buildAndBindSlickDatabases(dbApi, actorSystem.settings.config, applicationLifecycle)
     new SlickProvider(actorSystem)(executionContext)
   }
 }
